ass_voz_mult_idiom/requirements.txt
openai
openai-whisper
gTTS
sounddevice
scipy
python-dotenv
ipython
jupyter